Full-codebase audit (74 fixes) + status category flags on StatusItem (9.1.2) (#116)
* harness: add [3c.3] ENMITY TABLES section — always show both lists

The harness previously only surfaced enmity via TargetInfo.EnmityItems,
which Reader.Target.cs gates behind CurrentTargetID > 0 — so the table
was invisible when the player had no target. Two distinct enmity-related
structures exist on UIState and both deserve visibility for testing:

  AGRO  list (UIState.Hater._haters, HaterCount)
    AGROMAP_KEY / AGRO_COUNT_KEY — "who is aggro'd on the player"
    Already surfaced on PlayerInfo.EnmityItems; reprinted with the raw
    counter alongside for verification.

  HATE  list (UIState.Hate._hateInfo, HateArrayLength)
    ENMITYMAP_KEY / ENMITY_COUNT_KEY — "the player's own hate table"
    Read directly from memory using HateItem offsets so the data is
    visible regardless of whether GetTargetInfo would populate it.
    Names looked up against GetActors() since hate entries store only
    ID + Enmity.

Also logs CanGetAgroEntities / CanGetEnmityEntities capability flags so
a stripped scan (missing signature) is obvious at a glance, and lists
up to 10 actors with IsAgroed || InCombat set — independent of either
table — as an eye-check that ActorItem.IsAgroed lights up correctly.

The note above the TargetInfo.EnmityItems print is updated to point at
[3c.3] so it's clear why the same data shows up twice in different forms.

Harness-only — no Sharlayan source changes, no version bump.


* chore: bump FCS submodule 5d3840d68 → b21433ecc

Upstream rolled two commits since the previous pin:

  b21433ecc Add names from EventSystemDefine as comments
  e04ec6b37 Update structs

Routine struct refresh + a comments-only addition. No public-API,
field-rename, or layout-shape changes that Sharlayan would notice.
Lumina 7.5.0, Lumina.Excel 7.4.3, NLog 6.1.3, Newtonsoft.Json 13.0.4,
and ILRepack.Lib.MSBuild.Task 2.0.45 are already at latest stable
(verified via `dotnet list package --outdated` — no available updates).

Build clean, 189/189 tests pass, merged Sharlayan.dll byte size
unchanged at 10980 KB, debug DLLs deployed to Chromatics.


* fix: codebase audit — 69 verified bug/perf/memory fixes across readers, resolvers, and memory core

Multi-agent audit of the full Sharlayan codebase (FCS submodule excluded),
every finding adversarially verified before fixing. Highlights:

- ActorItem/PartyMember.Clone() now copy StatusNameEnglish (was silently null)
- GetEventObjectType receives the real actor address (was always IntPtr.Zero)
- Inventory container stride derived from FCS Size (was hardcoded 24 vs actual 32)
- Materia IDs read as ushort (Dawntrail grades > 255 no longer truncated)
- RPR/SGE levels + EXP wired end-to-end in PlayerInfo (were always 0)
- Party member Coordinate Y/Z typo; duplicate status adds; enmity NRE chain
- GetActors clones only genuinely-removed actors (was deep-cloning everything per poll)
- GetGameState derives IsLoggedIn from a pointer read (was running full player resolve)
- MemoryHandler primitive getters use thread-static buffers (zero alloc per read)
- Chat pipeline: bounds guards, compiled regexes, LINQ/copy churn removed
- Scanner: per-region exception flood removed, signature bytes hoisted out of scan loop
- P/Invoke hygiene: CloseHandle BOOL, GetStructure try/finally + correct handle
- DEPENDENCY.md + FCS integrity tests synced (InventoryContainer.SourceSize, RPR/SGE)

Build clean, 190/190 tests pass. Version 9.0.40.


* perf: resolve the five API-affecting audit findings (9.1.0)

- ActionResult.ActionContainers / InventoryResult.InventoryContainers:
  ConcurrentBag<T> -> List<T> (populated sequentially; the bag's
  thread-local machinery was pure overhead per poll)
- AstrologianResources.DrawnCards: List<AstrologianCard> -> AstrologianCard[]
- DancerResources.Steps: List<DanceStep> -> DanceStep[] (built directly,
  no temp array); CurrentStep adapted, Steps defaults to empty
- MonkResources.BeastChakra: stored array -> computed from BeastChakra1-3
  (zero per-poll allocation; values identical on access)
- InventoryItem materia arrays: shared zero-filled length-5 arrays for
  items with no melds; per-item arrays only when materia present
- ActorItemResolver: fallback names for nameless actors cached per type id

Minor version to 9.1 for the public API surface changes (authorized).
Build clean, 190/190 tests pass.


* chore: bump FCS submodule b21433ecc -> 15ae1806b (9.1.1)

38 upstream commits (struct updates, WeatherInterface, TerritoryInfo/
MapRangeFlags updates, EventFramework.SetDirectorData). Build clean,
190/190 tests pass including the FCS dependency integrity suite -
no Sharlayan-side changes needed.


* feat: expose StatusCategory / IsBeneficial / IsDetrimental / CanDispel on StatusItem (9.1.2)

Adds buff-vs-debuff classification to every status entry surfaced on
GetCurrentPlayer().Entity.StatusItems, GetActors(), GetTargetInfo(), and
GetPartyMembers(). Sourced from Lumina's Status sheet:

  StatusCategory  raw byte — 1 = beneficial (enhancement), 2 = detrimental
                  (enfeeblement), 0 = neither. Same discriminator the game
                  UI uses for the green/red status icon border.
  IsBeneficial    computed — StatusCategory == 1
  IsDetrimental   computed — StatusCategory == 2
  CanDispel       Esuna-removable (white-bar debuffs)

Unresolved status IDs fall back to StatusEffectLookup.DefaultStatusInfo,
which leaves StatusCategory at 0 — both convenience flags read false
rather than misclassifying garbage slots.

Plumbing: Models.XIVDatabase.StatusItem carries the two new columns,
LuminaXivDatabaseProvider populates them in the bulk Status pass, and
both ActorItemResolver and PartyMemberResolver copy them onto each
resolved Core.StatusItem alongside the existing name/stacks fields.
IStatusItem gains the four members (same additive pattern as
StatusNameEnglish in 9.0.x).

Harness [3c.1] is now a full STATUS EFFECTS section: per-entry
[B]/[D]/[-] tag plus all four new fields, dumped for the local player,
the current target (where DoTs/debuffs actually land — easiest
IsDetrimental eye-check), and up to three party members. Note the old
Duration <= 0 filter is gone, so permanent statuses (tank stances,
mount buffs) now appear with a resolved name.

4 new tests pin the category mapping (1=B, 2=D, 0/3+/255=neither).
195/195 pass. Version 9.1.1 → 9.1.2.


---------

+++ Sharlayan/Core/Enums/Inventory.cs
@@ -133,7 +133,7 @@ namespace Sharlayan.Core.Enums {
             //Unknown27008 = 27008,
         }
 
-        public enum MateriaType : byte {
+        public enum MateriaType : ushort {
             None = 0,
 
             Cracked,
